Update files to use absolute imports
Let's add all the parsing-related files to a subdirectory and use absolute imports.

This specific configuration for the CMakeLists files causes all files in the subdirectory to throw the warning "This file does not belong to any project target". More work will have to be done to rectify this soon.