Add new argumentToIndexMap field
Many methods, particularly findConnectingArgs and getColumnIndices, require finding the index of a particular column in a table or whether or not a certain argument is in the table. Each time these methods are called, a new unordered_set or unordered_map is created and populated, which is inefficient in the long term. Thus, ClauseResult will now also have an unordered_multimap<ClauseArgument, int> to keep track of what columns are in the table and their respective indices, alongside the existing vector<ClauseArgument>.